move search model to multiple files (#231164)
* move search model to multiple files to enable more diverse search tree
