+++ src/vs/workbench/contrib/search/test/browser/searchNotebookHelpers.test.ts
@@ -11,13 +11,16 @@ import { ICellViewModel } from '../../../notebook/browser/notebookBrowser.js';
 import { CellKind } from '../../../notebook/common/notebookCommon.js';
 import { contentMatchesToTextSearchMatches, webviewMatchesToTextSearchMatches } from '../../browser/notebookSearch/searchNotebookHelpers.js';
 import { CellFindMatchModel } from '../../../notebook/browser/contrib/find/findModel.js';
-import { CellMatch, FileMatch, FolderMatch, SearchModel, textSearchMatchesToNotebookMatches } from '../../browser/searchModel.js';
+import { SearchModelImpl } from '../../browser/searchTreeModel/searchModel.js';
 import { URI } from '../../../../../base/common/uri.js';
 import { TestInstantiationService } from '../../../../../platform/instantiation/test/common/instantiationServiceMock.js';
 import { createFileUriFromPathFromRoot, stubModelService, stubNotebookEditorService } from './searchTestCommon.js';
 import { IModelService } from '../../../../../editor/common/services/model.js';
 import { INotebookEditorService } from '../../../notebook/browser/services/notebookEditorService.js';
 import { ensureNoDisposablesAreLeakedInTestSuite } from '../../../../../base/test/common/utils.js';
+import { CellMatch, NotebookCompatibleFileMatch, textSearchMatchesToNotebookMatches } from '../../browser/notebookSearch/notebookSearchModel.js';
+import { FolderMatchImpl } from '../../browser/searchTreeModel/folderMatch.js';
+import { INotebookFileInstanceMatch } from '../../browser/notebookSearch/notebookSearchModelBase.js';
 
 suite('searchNotebookHelpers', () => {
 	let instantiationService: TestInstantiationService;
@@ -201,20 +204,20 @@ suite('searchNotebookHelpers', () => {
 		});
 
 
-		function aFileMatch(): FileMatch {
+		function aFileMatch(): INotebookFileInstanceMatch {
 			const rawMatch: IFileMatch = {
 				resource: URI.file('somepath' + ++counter),
 				results: []
 			};
 
-			const searchModel = instantiationService.createInstance(SearchModel);
+			const searchModel = instantiationService.createInstance(SearchModelImpl);
 			store.add(searchModel);
-			const folderMatch = instantiationService.createInstance(FolderMatch, URI.file('somepath'), '', 0, {
+			const folderMatch = instantiationService.createInstance(FolderMatchImpl, URI.file('somepath'), '', 0, {
 				type: QueryType.Text, folderQueries: [{ folder: createFileUriFromPathFromRoot() }], contentPattern: {
 					pattern: ''
 				}
 			}, searchModel.searchResult.plainTextSearchResult, searchModel.searchResult, null);
-			const fileMatch = instantiationService.createInstance(FileMatch, {
+			const fileMatch = instantiationService.createInstance(NotebookCompatibleFileMatch, {
 				pattern: ''
 			}, undefined, undefined, folderMatch, rawMatch, null, '');
 			fileMatch.createMatches(false);
